Move service setup extensions to extensions namespace and rename classes
